Refactored app.rb to controllers
