+++ app/controllers/peeps.rb
@@ -1,0 +1,6 @@
+class Chitter < Sinatra::Base
+  get '/peeps' do
+    @peeps = Peep.all
+    erb :'peeps/view'
+  end
+end
